Client, Server and Tests changed because of the strict flags

+++ src/client/client_SdlWindow.cpp
@@ -337,7 +337,7 @@ SdlWindow::~SdlWindow() {
       delete drawable;
     }
 
-    for (int i = 0; i < beams_textures.size(); i++){
+    for (size_t i = 0; i < beams_textures.size(); i++){
         delete beams_textures[i];
     }
